[gecko-camera] Use media buffers for decoded video. JB#56755
This commit introduces the GraphicBuffer abstraction for the raw video
data. The GraphicBuffer can be mmap'ed to YCbCrFrame or RawImageFrame
depending on the image format.

After applying this commit, the droid decoder will use the YUVFlexible
image format and media buffers to deliver the decoded data. This behavior
is enabled by default.

On some devices droid_media_buffer_lock() is slow, so the use of media
buffers should be disabled with GECKO_CAMERA_DROID_NO_MEDIA_BUFFER
environment variable.